# maas-api/openapi3.yaml
openapi: 3.0.3
info:
    title: Models as a Service API
    description: Models as a Service Billing and Management API
    version: "1.0"
servers:
    - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
    /v1/api-keys/bulk-revoke:
        post:
            tags:
                - api-keys-v2
            summary: Bulk revoke API keys for a user
            description: Revokes all active API keys for a specific user. Regular users can only bulk revoke their own keys. Admins can bulk revoke any user's keys.
            operationId: api-keys-v2#bulk-revoke
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required:
                                - username
                            properties:
                                username:
                                    type: string
                                    description: Username whose keys should be revoked
                        examples:
                            revoke_own:
                                summary: User revoking own keys
                                value:
                                    username: "alice"
                            admin_revoke:
                                summary: Admin revoking user's keys
                                value:
                                    username: "bob"
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    revokedCount:
                                        type: integer
                                        description: Number of keys that were revoked
                                    message:
                                        type: string
                                        description: Success message
                            example:
                                revokedCount: 5
                                message: "Successfully revoked 5 active API key(s) for user alice"
                "400":
                    description: Bad Request. Invalid request.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                "401":
                    description: Unauthorized response.
                "403":
                    description: Forbidden. User trying to revoke another user's keys.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                            example:
                                error: "Access denied: you can only bulk revoke your own API keys"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT  # optional, for documentation purposes only
  schemas:
...
        ErrorResponse:
            type: object
            properties:
                error:
                    type: string
                    description: Error message
                    example: Failed to retrieve models
            required:
                - error
        
        # Health check response
...